Initial commit first component and screen
On main.dart added themes ofr fonts and colors, and the home screen is categories_screen.dart.
categories_screen.dart have a grid layout that contains a category_item.dart component;
this component stylizes the itens in the dummy list that use the category.dart as model.

the fonts are defined in the pubspec.yaml.